floorplan tool/requirements.txt

pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
streamlit>=1.36.0
paho-mqtt>=2.1.0
requests>=2.31.0
plotly>=5.22.0
